pyproject.toml
[project]
name = "grab-api"
version = "0.0.1"
description = "grab-api"
dependencies = [
    'fastapi[standard]==0.114.0',
    'uvicorn==0.30.6',
    'SQLAlchemy==2.0.34',
    'python-dotenv==1.0.1',
    'alembic==1.13.2',
    'psycopg2==2.9.9',
    'pydantic==2.9.1',
    'PyJWT==2.9.0',
    'pytest==8.3.2',
    'platformdirs==4.3.2',
]
requires-python = ">=3.12"

[tool.setuptools]
py-modules = []

# Black formatter
[tool.black]
line-length = 79
